Chatbot_GenAI_Backend-main/requirements.txt
~flask
flask-cors
python-dotenv
pypdf2
python-docx
python-pptx
openpyxl
pymupdf
faiss-cpu
together
certifi
chromadb
numpy
huggingface-hub
sentence-transformers